requirements.txt
numpy
opencv-python
pandas
PyYAML
scikit-learn
scipy
torch==1.11.0
torchvision==0.12.0
tqdm
trimesh
wandb
monai
pyrender==0.1.33
pytorch3d==0.7.2